Added orbitlePulse animation
