+++ public/index.ts
@@ -6,3 +6,4 @@ export { default as CREATOR } from "./amitphoto.jpeg";
 export { default as MORPHINGBLOBSCENE } from "./morphingBlobScene.gif";
 export { default as DEVELOPERWORKSPACE } from "./developerWorkspace.gif";
 export { default as MYSTICORBITALS } from "./mysticOrbitals.gif";
+export { default as ORBITPULSE } from "./orbitPulse.gif";
